:recycle: Rename image util file, get contact from Sanity
